requirements.txt
absl-py==0.8.0
astor==0.8.0
backports.weakref==1.0.post1
certifi==2019.6.16
chardet==3.0.4
cloudpickle==1.2.2
contextlib2==0.5.5
decorator==4.4.0
dm-sonnet==1.34
enum34==1.1.6
funcsigs==1.0.2
futures==3.3.0
gast==0.3.2
grpcio==1.23.0
h5py==2.10.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.16.5
protobuf==3.9.1
requests==2.22.0
semantic-version==2.8.2
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
tensorflow-probability==0.5.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==0.15.6
wrapt==1.11.2
